requirements.txt
python-telegram-bot>=20.0,<21.0
urllib3
six
